requirements.txt
kafka-python
redis
paho-mqtt
